feat(registration): return 409 Conflict on duplicate and normalize age to int
Introduce RegistrationConflictException (extends RuntimeException, not IAE) so
that a duplicate email/username attempt returns HTTP 409 instead of 400.  The
exception carries a single generic message regardless of which field conflicts,
preventing enumeration of existing accounts through different error strings.

Wiring: caught in UserController before the broad IllegalArgumentException
catch, and in GlobalExceptionHandler as a safety-net fallback.

Also normalise User.age, IAuthenticationGateway.verifyUserDetails(), both
gateway implementations, UserResponse, and DevDataSeeder from float to int so
the domain model is type-correct and Mockito matchers (anyInt vs anyFloat) are
unambiguous.

7 new tests: conflict on email/username, generic message assertion, age-as-int,
and concurrent-registration tests updated to catch RegistrationConflictException.

+++ Version1/src/main/java/org/example/ApplicationLayer/IAuthenticationGateway.java
@@ -5,6 +5,6 @@ public interface IAuthenticationGateway {
     boolean verifyPassword(String plainPassword, String hashedPassword); // מיוז קייס 3
     boolean verifyEmail(String email);
     boolean verifyPassword(String pass);
-    boolean verifyUserDetails(String email, String password, float age, String username);
+    boolean verifyUserDetails(String email, String password, int age, String username);
 
 }
