+++ Version1/src/main/java/org/example/InfrastructureLayer/PlainTextAuthenticationGateway.java
@@ -36,7 +36,7 @@ public class PlainTextAuthenticationGateway implements IAuthenticationGateway {
     }
 
     @Override
-    public boolean verifyUserDetails(String email, String password, float age, String username) {
+    public boolean verifyUserDetails(String email, String password, int age, String username) {
         return verifyEmail(email)
                 && verifyPassword(password)
                 && age >= 0
